% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_aff_dict.R
\name{auth_aff_dict}
\alias{auth_aff_dict}
\title{Retrieve and recognize affiliation and country of affiliation from PMC PubMed Central}
\usage{
auth_aff_dict(authors_names, affiliation_dict)
}
\arguments{
\item{affiliation_dict}{List of affiliation copied and pasted from PMC PubMed Central}

\item{authors_name}{Authors list copied and pasted from PMC PubMed Central}
}
\value{
This function returns a \code{data.frame} with the list of authors of a queried scientific publication and detected country of affiliations
}
\description{
Creates a dictionary-like author-affiliation and dentifies country based on affiliation free-text field
}
\details{
This function was desinged to be applied to PubMed indexed studies published before 2014 (PubMed did not include all author affiliation information in their metadata) and for a few studies > 2014 with non-indexed affiliations.
}
\author{
Hugo Fitipaldi
}
