% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_university_state.R
\name{detecting_country_country}
\alias{detecting_country_country}
\title{Find country on affiliation field based on country name}
\usage{
detecting_country_country(x)
}
\arguments{
\item{x}{free-text (affiliation field)}
}
\value{
This function returns a character string with the name of the country
}
\description{
Matches country names on free text affiliation field
}
\details{
This function was designed to internal use in the package. The aim is to find a match of a country name on the free-text affiliation field.
}
\examples{
detecting_country_country("Lund University Diabetes Centre, Malmö, Skåne, Sweden")

}
\author{
Hugo Fitipaldi
}
