% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_university_state.R
\name{detecting_country_state}
\alias{detecting_country_state}
\title{Find country on affiliation field based on name of state}
\usage{
detecting_country_state(x)
}
\arguments{
\item{x}{free-text (affiliation field)}
}
\value{
This function returns a character string with the name of the country
}
\description{
Find country name on free text affiliation field based on state name
}
\details{
This function was designed to internal use in the package. The aim is to find the country name by matching with a name of state.
}
\examples{
detecting_country_state("Lund University Diabetes Centre, Malmö, Skåne, Sweden")

}
\author{
Hugo Fitipaldi
}
