% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_university_state.R
\name{detecting_country_university}
\alias{detecting_country_university}
\title{Find country on affiliation field based on university name}
\usage{
detecting_country_university(x)
}
\arguments{
\item{x}{free-text (affiliation field)}
}
\value{
This function returns a character string with the name of the country
}
\description{
Find country names on free text affiliation field based on university names
}
\details{
This function was designed to internal use in the package. The aim is to find the country name by matchin with the university location.
}
\examples{
detecting_country_university("Lund University Diabetes Centre, Malmö, Skåne, Sweden")

}
\author{
Hugo Fitipaldi
}
