% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_university_state.R
\name{detecting_us_states}
\alias{detecting_us_states}
\title{Detect US states abbreviates}
\usage{
detecting_us_states(x)
}
\arguments{
\item{x}{free-text (affiliation field)}
}
\value{
This function returns a character string with the name of the detected US state.
}
\description{
Finds US state abbreviations.
}
\details{
This function was designed to detect US state abbreviations. Internal use only.
}
\examples{
detecting_us_states("OH is a United Stated state.")

}
\author{
Hugo Fitipaldi
}
