% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{get_geodata_from_json_nominatim}
\alias{get_geodata_from_json_nominatim}
\title{Pre-processing of json transformed query}
\usage{
get_geodata_from_json_nominatim(geodata_json)
}
\arguments{
\item{geodata_json}{}
}
\value{

}
\description{
Matches country names on free text affiliation field
}
\details{
This function was designed to internal use in the package. The aim is to find a match of a country name on the free-text affiliation field.
}
\author{
Hugo Fitipaldi
}
