% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_university_state.R
\name{replace_us_state}
\alias{replace_us_state}
\title{Detect and replace US states abbreviations to full state name.}
\usage{
replace_us_state(x)
}
\arguments{
\item{x}{free-text (affiliation field)}
}
\value{

}
\description{
Finds US state abbreviations and replace them to the states full name within a text.
}
\details{
This function was designed to replace US states abbreviations to its full name within a text.
}
\examples{
replace_us_state("OH is a United Stated state.")

}
\author{
Hugo Fitipaldi
}
