% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_functions.R
\name{url_nominatim_search}
\alias{url_nominatim_search}
\title{Matches country names on free text affiliation field}
\usage{
url_nominatim_search(search_query_url, country_url, language_url, email_url)
}
\arguments{
\item{email_url}{}
}
\value{
This function returns a character string with the name of the country
}
\description{
This function was designed to internal use in the package. The aim is to find a match of a country name on the free-text affiliation field.
}
\author{
Hugo Fitipaldi
}
