% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affiliations.R
\name{get_affiliations}
\alias{get_affiliations}
\title{Retrieve and recognize affiliation and country of affiliation from PubMed}
\usage{
get_affiliations(query, email = NULL, format.long = FALSE)
}
\arguments{
\item{query}{selects which data set from CSSS you want to download: "national" (default), "county" or "postcode"}

\item{format.long}{}
}
\value{
This function returns a \code{data.frame}
}
\description{
Identifies country based on affiliation free text field
}
\details{
This function was designed to
}
\examples{
get_affiliations(query = '30237159')
get_affiliations(query = '30237159', format.long = TRUE)

}
\author{
Hugo Fitipaldi
}
